pyproject.toml
[project]
name = "issue-tracker-api"
version = "0.1.0"
description = "A Flask-based issue tracker API"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "alembic==1.16.2",
    "annotated-types==0.7.0",
    "attrs==25.3.0",
    "blinker==1.9.0",
    "click==8.2.1",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "flasgger==0.9.7.1",
    "flask==3.1.1",
    "flask-jwt-extended==4.7.1",
    "flask-migrate==4.1.0",
    "flask-sqlalchemy==3.1.1",
    "greenlet==3.2.3",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "mistune==3.1.3",
    "packaging==25.0",
    "psycopg2==2.9.10",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyjwt==2.10.1",
    "python-dotenv==1.1.1",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "rpds-py==0.25.1",
    "six==1.17.0",
    "sqlalchemy==2.0.41",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "werkzeug==3.1.3",
    "gunicorn>=21.2.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-flask",
    "coverage",
    "ruff",
    "black",
    "pre-commit"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = ["migrations", ".venv", ".env", "env", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I"]
fixable = ["ALL"]